update packaging to start with 'com.commentremover' update version to 1.2-SNAPSHOT add Apache 2.0 license remove '/src/test' from .gitignore
