+++ .gitignore
@@ -66,5 +66,3 @@ crashlytics-build.properties
 /target
 
 /resources
-
-/src/test
