+++ pom.xml
@@ -6,7 +6,7 @@
 
     <groupId>com.commentremover</groupId>
     <artifactId>comment-remover</artifactId>
-    <version>1.0</version>
+    <version>1.2-SNAPSHOT</version>
     <packaging>jar</packaging>
 
     <properties>
@@ -54,8 +54,6 @@
                 </plugin>
             </plugins>
         </pluginManagement>
-
     </build>
-
 
 </project>
