+++ src/main/java/com/commentremover/processors/AbstractFileProcessor.java
@@ -1,7 +1,7 @@
-package processors;
+package com.commentremover.processors;
 
-import app.CommentRemover;
-import exception.CommentRemoverException;
+import com.commentremover.app.CommentRemover;
+import com.commentremover.exception.CommentRemoverException;
 
 import java.io.BufferedReader;
 import java.io.BufferedWriter;
