+++ src/main/java/com/commentremover/processors/CSSFileProcessor.java
@@ -1,19 +1,19 @@
-package processors;
+package com.commentremover.processors;
 
-import app.CommentRemover;
-import exception.CommentRemoverException;
-import handling.RegexSelector;
+import com.commentremover.app.CommentRemover;
+import com.commentremover.exception.CommentRemoverException;
+import com.commentremover.handling.RegexSelector;
 
 import java.io.File;
 import java.io.IOException;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-public class XMLFileProcessor extends AbstractFileProcessor {
+public class CSSFileProcessor extends AbstractFileProcessor {
 
-    private static final String regex = RegexSelector.getRegexByFileType("xml");
+    private static final String regex = RegexSelector.getRegexByFileType("css");
 
-    public XMLFileProcessor(CommentRemover commentRemover) {
+    public CSSFileProcessor(CommentRemover commentRemover) {
         super(commentRemover);
     }
 
@@ -29,12 +29,17 @@ public class XMLFileProcessor extends AbstractFileProcessor {
 
     @Override
     protected StringBuilder doRemoveOperation(StringBuilder fileContent, Matcher matcher) throws StackOverflowError{
+
         String sFileContent = fileContent.toString();
         boolean isTodosRemoving = commentRemover.isRemoveTodos();
 
         while (matcher.find()) {
 
             String foundToken = matcher.group();
+
+            if (isDoubleOrSingleQuoteToken(foundToken)) {
+                continue;
+            }
 
             if (isTodosRemoving) {
                 sFileContent = sFileContent.replaceFirst(Pattern.quote(foundToken), "");
@@ -44,8 +49,14 @@ public class XMLFileProcessor extends AbstractFileProcessor {
                 }
             }
         }
+
         fileContent = new StringBuilder(sFileContent);
 
+
         return fileContent;
+    }
+
+    private boolean isDoubleOrSingleQuoteToken(String foundToken) {
+        return foundToken.startsWith("\"") || foundToken.startsWith("\'");
     }
 }
